feat: 首页功能 1、商品详情 2、商品评论 3、商品搜索 4、购物车

+++ foodie-common/src/main/java/com/lpf/utils/PagedGridResult.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class PagedGridResult {
 	
 	private int page;			// 当前页数
-	private int total;			// 总页数	
+	private long total;			// 总页数
 	private long records;		// 总记录数
 	private List<?> rows;		// 每行显示的内容
 
@@ -19,10 +19,10 @@ public class PagedGridResult {
 	public void setPage(int page) {
 		this.page = page;
 	}
-	public int getTotal() {
+	public long getTotal() {
 		return total;
 	}
-	public void setTotal(int total) {
+	public void setTotal(long total) {
 		this.total = total;
 	}
 	public long getRecords() {
